// client/src/components/Artist/ArtistFormInterfaces.ts
export interface ArtistFormValues {
    artistName: string;
    image: string;
    role: string[];
    bio: string;
    twitter_link: string;
    instagram_link: string;
    facebook_link: string;
    soundcloud_link: string;
    bandcamp_link: string;
    youtube_link: string;
    spotify_link: string;
    apple_music_link: string;
    beatport_link: string;
  }